+++ src/Form/UserActivityType.php
@@ -1,11 +1,12 @@
 <?php
 
+
 namespace App\Form;
 
 use App\Entity\UserActivity;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -13,20 +14,23 @@ class UserActivityType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('activities', CollectionType::class, [
-            'entry_type' => ActivityType::class,
-            'entry_options' => ['label' => false],
-        ]);
-        $builder->add('Enregistrer', SubmitType::class, [
-            'attr' => ['class' => 'btn btn-primary row ml-2 mb-3'],
-        ]);
+        $builder
+            ->add('hour', IntegerType::class, [
+                'label' => 'Heures',
+                'empty_data' => '0',
+                'error_bubbling' => true,
+            ])
+            ->add('minute', IntegerType::class, [
+                'label' => 'Minutes',
+                'empty_data' => '0',
+                'error_bubbling' => true,
+            ]);
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
             'data_class' => UserActivity::class,
-            'label' => false,
         ]);
     }
 }
